Added logo, header and footer colour and text on the hero section.
Changed the small top left picture with a logo and a text on its side, with the navigation bar on the right with a drop down menu. And changed the text on the index page to appear on top of the hero image, and the footer to have the same colour as the header.